requirements.txt
fastapi
uvicorn
pywebview
pyinstaller